Add authorization to modify listings

+++ app/Http/Controllers/ListingController.php
@@ -51,6 +51,10 @@ class ListingController extends Controller
     }
 
     public function update(Request $request, Listing $listing){
+        if($listing->user_id != auth()->user()->id){
+            abort(403,'Unauthorized action');
+        }
+
         $formFields  = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -70,8 +74,14 @@ class ListingController extends Controller
     }
 
     public function destroy(Listing $listing){
+        if($listing->user_id != auth()->user()->id){
+            abort(403,'Unauthorized action');
+        }
         $listing->delete();
 
         return redirect("/")->with('message', 'Listing deleted successfully');
+    }
+    public function manage(){
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
